requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
tzdata==2025.2
Pillow==10.4.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
requests==2.32.3
whitenoise==6.7.0
